Add: fiters-folder. All custom filters moved there.
RouteViewSet has RouteFilter from filters/filters.py.
